Replace example2input with form2input
form2input gets the entire form dict as input, so that it can pull other
data than the example from it if necessary. A future commit will use
this to re-populate the form with the user’s input (i. e., set the value
attribute) when asking them to confirm that their lexeme is not a
duplicate of an existing one.