Add advanced mode
In advanced mode, all forms are optional, which can be useful for
certain words (e. g. singularia tantum or pluralia tantum).

The template gains some additional elements, mostly dependent on the
presence or absence of `advanced` in the context. The form2input filter
also needs the context for this, so it is decorated as a context filter.

(As in a previous commit, we explicitly add jinja2 to requirements.txt
now that we directly import it, even though it was clearly already
required earlier and is pulled in by flask anyways.)