Group templates on index page
Currently, they are only grouped by the language code, but the grouping
mechanism is kept in a function in app.py, so it hopefully shouldn’t be
too hard to improve it in the future.

Note that we don’t sort the templates according to their group, so if
there are discontiguous templates with the same group in the
templates.py OrderedDict, they will appear separately on the index page.