Detect duplicate lexemes client-side
We do an API request each time the user “commits” a change to the first
form representation (i. e., it loses focus), and add the results to the
DOM (after first removing the corresponding, now stale elements). For
users with JavaScript disabled, nothing should change – the tool still
detects duplicates server-side as well if the corresponding checkbox
isn’t checked.

Note that we only listen to changes on the first form representation,
which is not necessarily the one that will end up as the lemma in
advanced mode. I think that’s an acceptable limitation. (Server-side
duplicate detection checks the actual lemma, so it should still work,
just not as conveniently.)