Add template for English nouns
English nouns are fairly simple, but this is still useful because of the
duplicate detection, and so people can try out the tool in a language
they likely understand even if they don’t speak any of the other
languages the tool supports.

Forms for possessive case could be added in the future, but for now it’s
simpler to leave them out.