Also check for duplicates after 500 ms of no input
Use lodash’s debounce function to check for duplicates after the user
has not typed anything for 0.5 seconds, even if they haven’t committed
the change yet to trigger a 'change' event. Use the “flush”
functionality to still check for duplicates immediately on commited
changes to the field.

Suggested by @addshore on Twitter.