Preserve form data when switching to advanced mode
To avoid throwing away all user input when switching to advanced mode,
turn the simple link into a secondary button that POSTs to the advanced
page, and make the advanced page only process a POST when the referrer
matches that page. (I suppose this also doubles as a sort of CSRF
protection, though the CSRF token should take care of that already.)
Note that this will not work if a Referrer-Policy [1] of no-referrer is
specified, which is fine because Toolforge doesn’t send any
Referrer-Policy by default and it’s not the browser default either.

This improvement was suggested by [[User:Nikki]] [2] – thanks!

[1]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Referrer-Policy
[2]: https://www.wikidata.org/wiki/User:Nikki