Remove workaround for T194732
T194732 is fixed, and the fix is deployed to both test- and real
wikidata now (and it doesn’t look like it will be rolled back again), so
we can get rid of this unfortunate workaround now and rely on the API to
do the right thing by itself. Yay!