Allow messages to fall back to English
I’m about to add a new feature which requires a new message, and I can’t
wait for the translators to provide translations for the message before
adding the feature (could take a while), so the app needs to be able to
handle missing translations. This implements the simplest version: a
fallback to the English version, which should always be available.
Perhaps I’ll add more sophisticated fallbacks between related languages
in the future.