Add hint for switching to advanced mode
If users try to submit the form with some forms missing, show them a
hint that what they seem to be trying to do is only available in
advanced mode.

The implementation is similar to the client-side duplicates warning: the
server offers an API returning HTML (in this case more static than for
the duplicate search), and the client calls that API and injects the
result into the page when appropriate. In this case, we don’t bother
with ever removing the hint, since it’s not necessary.

The hint message is only available in English for now – I’ll ask the
community to translate it into the other languages.