Avoid breaks between form label and definition
If the list of forms is distributed across several columns, the browser
may place a break between the label of a form and its definition. To
avoid this, wrap them both in a block element and instruct the browser
to not break inside this element.

This currently doesn’t have any effect, since we don’t distribute the
form list across columns, but I’m thinking of adding that (see also [1])
and don’t want to leave this work-in-progress part of it sitting around
on my PC uncommitted for too long.

[1]: https://twitter.com/LucasWerkmeistr/status/1057409761929580545