Support storing OrderedMultiDict in session
Augment the session’s serializer so that it can completely serialize and
deserialize ordered multi-dicts (instead of only serializing the first
value per key and deserializing as a plain dict). This will be necessary
for an upcoming feature, but is also nice to have in general.
Implementation based on the Flask documentation at [1].

[1]: http://flask.pocoo.org/docs/1.0/api/#tagged-json