Insert “no duplicate” input before submit button
Instead of appending it after the last form representation, prepend it
before the submit button. I think this makes sense in general, but in
the particular case I have in mind right now where I want to add another
input, that input should be between the form representations and the
checkbox, instead of the checkbox being before it.