Add “generated via” form parameter
If there is a generated_via form parameter while building the form, add
it as a readonly input; if it is there on submit, append it to the
summary. This never comes into play during normal use, but will be
useful with an upcoming feature to initialize the form parameters from
URL query parameters.

Note that the “generated via” is translated on the form page, but not in
the actual summary we submit. This matches that the summary otherwise
consists of the template name, which is also in English.