Initialize form data from URL query parameters
If the tool is loaded with form parameters in the URL – for example,
/template/english-noun/?form_representation=dog&form_representation=dogs&via=something%20else –
then we stash the form data in the session, redirect the user to the
template page without the query parameters, and there use the stashed
form data to initialize the form data if it is not otherwise present.
This allows external tools to use this tool for auto-generating lexeme
forms: they generate all the form representations, put them in the URL
and then send the user there, preferably identifying themselves in the
generated_via parameter. The user can still check if the forms are
correct before submitting them. The advantage for the external tool is
that it doesn’t need to handle OAuth, grammatical feature item IDs,
duplicate lexemes or various other complications, and also that there
still is a layer of human verification before the user submits the form.

The reason we stash the form data and redirect the user is twofold:
first, I think it’s a bit ugly to leave all the parameters in the URL,
especially if the user edited the form representations so that the form
data no longer matches the query parameters; and second, current_url()
currently doesn’t take query parameters into account, so POSTing to the
page with any query parameters actually fails CSRF protection because
the referrer doesn’t appear to match the current URL.