Fully rename “via” to “generated_via”
While developing, I initially called this parameter just “via” and then
decided to rename it to “generated_via” for increased clarity and
consistency with the message. But apparently I didn’t do that in the
part that propagates the parameter from the URL to the form parameters,
so let’s fix that.