Use PluralFormatter for forms/senses message
Example:

> dog (L1122), English, noun, 2 forms and 1 sense

This requires the addition of a message_with_language() method so that
we can use the pluralization rules of the correct language even when
language fallbacks were applied.

Unfortunately, this doesn’t compose perfectly with MarkupSafe: we plug a
Markup object into the formatter but get a str instead of another Markup
out of it. Fortunately, that’s the safe direction (everything gets
escaped when sent back to Jinja2), and we don’t actually need any
unescaped markup in the text for now. But perhaps I’ll investigate how
to improve this later.