Allow explicit keys in PluralFormatter
MediaWiki has supported this since version 1.22 (background [1]), and
now we support it as well.

Also, I noticed that the class already supports substitutions in the
texts (because StringFormatter applies vformat to the format spec). I
thought I’d have to implement that explicitly, but this is even better,
so the documentation for that is also included in this commit.

Together, this results in a closer correspondence to MediaWiki’s syntax:

> {{PLURAL:$1|0=no changes|one change|$1 changes}} # MediaWiki
> {count} {count!q:one=change:other=changes} # Python, previously
> {count!q:0=no changes:one=one change:other={count} changes} # Python, now

(Note that this example would format -1 as “one change” in both the
MediaWiki and the second Python version.)

[1]: https://laxstrom.name/blag/2012/10/29/muir-woods-has-one-tree-plural-issues-in-mediawiki/