Fix JSON version of duplicates API
In 455b66909d, I changed get_duplicates() to have an OrderedDict and
return its .values(), rather than having a list and returning that
directly. I assumed that OrderedDict.values() returns a list, but in
fact it returns an odict_values object, which is very similar but
different enough that the JSON serializer doesn’t accept it.

None of the current tests could have caught this, because they
monkeypatch get_duplicates() to return a real list. Perhaps I’ll improve
that, but I don’t want to block deploying the fix on this, so for now
this commit doesn’t include a regression test for the bug it fixes.

Reported by User:Vesihiisi, who tried to use the API in their own tool.