Add API to match lexemes to templates
This broadly answers the question “which template(s) match(es) this
lexeme” and will be used by a user script to automatically add links to
advanced mode on lexemes with no forms. I’ve tried to make it general
enough that it may also be useful to others (in that it returns not just
one or several candidates but the match quality for all templates, and
leaves it to consumers such as the user script to decide how to weigh
the different components of the match).

The missing_statements part can also be used for a future feature:
currently, when you edit a lexeme in advanced mode missing statements on
the lexeme are not added, but I’d like to fix that. (For example, if
you’ve used a “feminine” template to edit a lexeme, add the
corresponding “grammatical feature” statement there.)