Fix and add tests for adding missing statements
I forgot to run the tests with the previous commit. This fixes the two
tests that it broke (adding empty 'claims' to the expected value, and
also monkeypatching get_lexeme_data() so the test doesn’t fetch data
from Wikidata), and adds another test for the new functionality.

I’m adding this as a separate commit, rather than amending the previous
one, because I already logged a message [1] referring to the old
commit’s hash.

[1]: https://wikitech.wikimedia.org/wiki/Special:Diff/1831655