Add parse_tpsv module
The syntax is the same as in QuickCategories, but with no special
interpretation of the fields – they’re just a list of form
representations, at least for now. (I might want to add support for
optional lexeme IDs as the first field later.) This will be used for a
new mode to create lexemes in bulk.