Distinguish between (non-)actionable duplicate warnings
This warning will also be displayed in the output of lexeme bulk
creation mode, where it won’t be actionable; therefore, it doesn’t need
an ID (there is no JavaScript to add/remove it dynamically, and the ID
wouldn’t be unique anyways), and the instructions at the bottom do not
apply. For now, though, all callers specify actionable=True.