Add bulk lexeme creation mode
In this mode, you can create a set of lexemes by specifying their form
representations, separated by tab or pipe characters, one lexeme per
line – for example:

    thing|things
    example|examples
    test|tests

This is deliberately not linked anywhere yet – I want to keep the
feature hidden and test it with some selected users initially. I’m not
even sure about the “bulk” name yet. (I hesitate to call it a “batch”,
as in QuickStatements or QuickCategories, because it’s not permanent and
I have no intention to make it so.)