Limit bulk mode to autoconfirmed users
This is a powerful tool that, for now, I don’t want to give to just
anyone. (can_use_bulk_mode() will also be used to decide whether to show
the link to bulk mode in the first place, once that link exists
anywhere, so normally unprivileged users shouldn’t see that error.)

When OAuth is not configured (i. e. in local development mode), anyone
is allowed to use bulk mode, so that it can still be worked on.