Distinguish Czech adjectives by comparison
Add the grammatical feature “positive” to all the current ones, and then
add sets of forms for “comparative” and “superlative”. This was done by
[[User:Adrijaned]] [1] in [[Special:Diff/1033244170]] [2] – thank you!

The existing forms were updated with the following sed command:

    sed -i "/czech-adjective/,/^    }/ { s/^\( \{16\}'label'.*\)',$/\1, 1. stupeň',/; s/^\( \{16\}'grammatical_features_item_ids'.*\)],$/\1, 'Q3482678'],/;}" templates.py

I also sanity-checked the transcription of the new forms by, after the
initial version of this commit, removing them and instead inserting two
copies of the first set of forms (positive), then replacing “1. stupeň”
with “2. stupeň” / “3. stupeň” and “Q3482678” with “Q14169499” /
“Q1817208”. The expected outcome of this was that `git diff` would only
show minor differences in the examples, confirming the regularity of the
labels and grammatical features – but in fact I had copied the label
instead of the example for two forms, which this made obvious. (Though
that kind of mistake could also be spotted by a test, seeing as the
label lacked the [placeholder] required for an example – but I don’t
seem to have implemented such a test yet.)

[1]: https://www.wikidata.org/wiki/User:Adrijaned
[2]: https://www.wikidata.org/wiki/Special:Diff/1033244170