Support specifying lexeme IDs in bulk mode
You can now specify the lexeme ID as an extra field at the beginning of
a line:

    form 1|form 2|form 3
    L123|form 1|form 2|form 3

Technically, this is unambiguous by the number of forms alone; however,
to guard against off-by-one errors, we also report an error if it looks
like a form is missing (i. e. what we would interpret as the first form
representation looks like a lexeme ID):

    L123|form 1|form 2

To implement this, the form data generation moves from app.py to
parse_tpsv.py, which unfortunately complicates the tests rather a bit.
As a result, several tests for parse_lexemes are combined into one.