Improve directionality on index page
We should set the dir= attribute wherever we set the lang= attribute.
But if we just add a dir= on the <li>, which previously had the lang=,
then the whole list item block is RTL, with the list marker all the way
to the right of the page, which looks odd. Instead, only mark the
*contents* of the list item RTL using a new surrounding <span> – that
way, the contents are in the right order (e. g. “bulk mode” to the left
of the template label) but the list markers remain consistent.