Correctly pluralize warning for duplicate lexemes
I just happened to notice this while looking at other messages (to be
added) – now that we have plural support (for the “x forms and y senses”
feature), it’s not necessary for the messages surrounding the duplicates
to always refer to the duplicates in plural number, they can also use
the correct plural form. Add a new template function to enable this.

When I tested this with the Latin translation, I discovered that it
crashed, because CLDR has no Latin locale. Presumably, duplicates search
for Latin templates has been broken since plural support was introduced;
fix this for now by mapping Latin to English (they have the same plural
distinction: one/other; Babel does not support providing custom locales,
so mapping to another language code is the only option as far as I’m
aware) and add a test that each template language can be pluralized for
several numbers (which should also cover the case where a language is
missing some plural form, at least for those numbers and in the one
message that’s used for the test).