Enable get_lexeme_data() to get data for revision ID
For the lexeme editing feature, we’ll need to load the data for a
specific revision ID. We can get that from Special:EntityData
(wbgetentities has no corresponding parameter).

In fact it would be better to use Special:EntityData for the latest
entity data as well – we might hit a cache and get a response more
quickly – but currently those cached results aren’t purged when an
entity is edited, which means that we might get stale data from there.