Add first version of edit mode
In this mode, we load the forms of an existing lexeme, match them to the
template forms, and then apply whatever changes the user makes to the
lexeme forms: edit form representations, delete or add forms. Lexeme
forms that can’t be unambiguously matched to a template form are listed
separately, the user is expected to edit those by hand if needed.

This mode isn’t linked to anywhere else yet, since I’d like to get some
feedback on it first. (In fact, I’m not sure where it would make sense
to link to it at all… we’ll see.)

There are no tests yet, but I hope to add some soon.