Simplify language codes for Babel
Babel doesn’t understand the -x-Qid stuff (e. g. mis-x-Q401) that we use
in Wikidata, and throws an error when it encounters it (a different one
than the error we catch). Trim that part away.

Note that this is distinct from the problem of the separator for the
locale identifier. By default, Babel expects the identifier to look like
en_US or zh_Hans_CN; it can be configured to accept en-US or zh-Hans-CN
instead (or we could easily do the replacement ourselves), but after
that we’re still left with the same problem that the -x-Qid stuff is not
the kind of locale information Babel expects. So I believe throwing it
away is actually the right solution.