Add test for match_lexeme_forms_to_template()
Some assertions are written as

    assert dict.get(key) is None

rather than the more intuitive

    assert key not in dict

because I noticed that pytest does not provide useful output for the
`not in` syntax: it prints the whole right-hand side (dict), abbreviated
(with some entries replaced by an ellipsis), so the offending data for
the key may not even be shown. The get() version should show the data
for the key if it exists.