Fix matching for english-adverb template
The english-adverb template has only a single form with no grammatical
features and no statements, which meant that when trying to edit an
existing lexeme, its form would be reported as not matching any form in
the lexeme. Accommodate this only for such special cases: if there is
more than one form, then zero-feature matches should not be allowed.

This corner case was uncovered by an integration test for the “edit”
feature that will be committed separately in the next commit; this
commit adds a unit test for just the corner case.