Add integration test for “edit” feature
This simulates creating a new lexeme, then trying to edit it – the
tool should correctly match the lexeme forms with the template forms.
One use of this is to check that the forms within each template can be
distinguished from one another. I was initially going to test this in
test_templates, checking that the grammatical features are always
distinct, but that is actually not always the case: the template
polish-noun-masculine-personal-with-depreciative-forms has forms with
the same set of grammatical features that only differ in their
statements. And I didn’t want to add logic to compare statements in the
templates unit test, so instead I turned this into an integration test
where the existing matching logic does the work and we only sanity-check
its results at the end.