Configure larger uWSGI buffer
The stashed_form_data mechanism can result in storing a lot of data in
the session, and the default buffer size for all request headers isn’t
enough to store the maximum cookie size that many browsers support (at
least not in conjunction with other request headers like User-Agent or
Host). Copied from SpeedPatrolling commit fcb85ef502.

In the long term we should probably try to reduce the amount of data we
store in the session (e. g. rename the absurd “form_representation” to
something shorter), but this is a relatively simple stop-gap solution
that should hopefully help at least a little.