Fix edge case bug in update_lexeme
In Python, removing elements from a list while iterating over it will
silently skip elements. (Java is friendlier, it throws a
ConcurrentModificationException in such cases.) This could lead to
matching lexeme and template forms not being recognized as matching,
causing them to be incorrectly modified. Fix that, and add a test (the
“noop” one: if the lexemes are not recognized as all matching, the form
representations will be altered because the lexeme forms and user input
are not in the same order), along with a more general test for
update_lexeme which is unaffected to this bug but still nice to have.
(The “iterate backwards” trick comes from StackOverflow [1].)

[1]: https://stackoverflow.com/a/1207485/1420237