Support manually matching forms in edit mode
The user can now specify a form ID in place of a form representation,
and that lexeme form will be made to match the template form, adding
missing grammatical features and statements. (If there are conflicting
statements, an error is raised.) The intended use is to “fix” forms that
were reported as unmatched, or possibly ambiguous (not sure about this);
that said, we don’t check this at the moment, and if the user specifies
the IDs of already matched forms, funny things will probably happen.

Currently, the ID has to be entered manually; the next commit will make
this easier to use.