Allow drag’n’drop of unmatched forms in edit mode
The user can now drag a URL from the list of unmatched (or ambiguous, I
guess) forms into the right input field, and a little JavaScript snippet
will correctly mangle it so that the backend recognizes it (form ID, not
full URL). As a bonus, we afterwards remove the dragged list entry as
well, to prevent accidentally dragging it a second time. Also included
is a little bit of CSS to make it more obvious where the user can drop
the links.