Disallow re-matching matched forms in edit mode
If a form is not in the list of forms that could not be matched for the
template, refuse to re-match it to a different template form. By
removing re-matched forms from that list, we additionally prevent the
case of trying to match the same form ID to several template forms at
the same time.