Add CommaSeparatedListFormatter
This will be used to format the list of grammatical features of a form,
when showing it as unmatched/ambiguous in edit mode. Fortunately, Babel
can do most of the work for us here.

Additionally, introduce I18nFormatter, which combines PluralFormatter
and CommaSeparatedListFormatter (and any future related formatters, if
we get more), supporting all their conversions.