Add ability to split template into sections
This allows a template to declare that it has multiple sections of
forms, and where the section breaks should go. The sections are
presented separated by Thematic Break (<hr>) elements, and within a
section, forms may be distributed across two columns. This should
hopefully make the tool more convenient to use, especially for longer
templates. Credit for the section idea goes to [[User:Nikki]] [1].

The way the sections are declared – using a section_break flag on an
individual form – is conceptually not very nice; however, the nicer
approach (where the forms would be a list of lists, each list
representing one section) would be much harder to make work with other
features, especially edit mode, which expect the forms to be a flat
list. I think the flag solution is acceptable here to keep the
maintenance cost manageable: this way, any part of the tool that’s not
aware of the sections can just ignore that flag and still work with the
list of forms just as before.

For now, only the German templates get sections added, but I plan to add
sections to other languages too – I should be able to determine good
places for section breaks, based on the grammatical features of the
forms, even for languages that I don’t speak.

[1]: https://www.wikidata.org/wiki/User:Nikki