Link to edit mode in duplicates warning
The duplicates template filter and API get an optional template_name
parameter, which is used to form a link to edit mode in a warning about
a potential duplicate lexeme. The places in the tool which warn about
duplicates all add this parameter as well. The message for the edit link
is the same as the regular edit message, except in initial lowercase
(converted by Emacs’ M-l downcase-word command).