Improve test_examples_distinct
Print all ambiguous examples at once, and also assert a specific, sharp
count of occurrences for each, not just “these examples are allowed to
occur as often as they like”. The count can be cross-checked against the
number of matches in the browser when searching for the example.

The “Nosotros [cantamos] mucho.” entry had become unnecessary with
commit 242c25810b, which disambiguated the example, I just didn’t think
to remove it from the tests at the time.