Make normal→bulk navigation more convenient
If the user is moving to bulk mode from normal mode, but had not yet
typed anything into the normal mode inputs, then the bulk mode input
will be pre-filled with just a bunch of empty pipes. This is not very
useful, and also hides the (potentially useful) placeholder. To prevent
this, clear the bulk mode value in this case.