Fix dragend event not being fired
It looks like recent browser versions (at least Firefox 82) no longer
fire the dragend event if the drag’n’drop was successful, so the dashed
border would only be removed if the drop ended outside a text area. Fix
this by calling the dragend handler in the drop handler (with the wrong
event but it doesn’t use the event anyways, and it’s also idempotent so
even if it does get called by dropend too it’s no problem).