Use correct first form for lemma
The lemma in edit mode should be based on the first form in template
order, not in lexeme order – in other words, the first of the lexeme
forms corresponding to the first template form, not the first lexeme
form. Test this by swapping müssen and muss and checking that muß is
still not used for the lemma even though it’s the first lexeme form.