Support gender-dependent messages
This is needed in the Italian bulk_not_allowed message – it depends on
the gender of the viewing user. Add a new message formatter which takes
a pluggable get_gender function and implement that function in app.py.
The function returns three specifiers: “m” for the grammatically male
version of the message, “f” for the grammatically female one, and “n”
for the grammatically neutral one. (I don’t mean this to imply that
there are only three genders – it’s only about the number of
distinctions made in supported languages. If another language comes
along with more gender distinctions, we can probably support that –
though MediaWiki doesn’t have any more fine-grained information for a
user name yet.)