Add language code back to index page
The language code can be useful, e.g. to search for a language whose
script you can’t read (when demoing the tool to people who speak the
language). It’s inside the outer <span> for the autonym because if the
autonym is RTL, the language code should be to the left of it; the
language code used, zxx, means “no linguistic content”, which seems
appropriate enough here.