Allow sections breaks without two-column layout
This renames the has_sections flag to two_column_sections and the
with-sections class to two-column-sections, to more accurately reflect
their purpose: not whether a template has sections (all templates do, at
least one) or section breaks (because templates without sections could
still declare the flag), but whether those sections should be arranged
in two columns or not. Also, it’s now allowed to set the flag to False,
to have section breaks without arranging the sections in two columns;
this will be useful for the Esperanto verb template.