Don’t use button to link edit mode from duplicates
If the link to the duplicates is a button, then it becomes the default
button of the whole form, and pressing Enter will go to edit mode, even
if the user actually wanted to submit the form as a duplicate lexeme.
Avoid this by turning the button back into a link; the next commit will
then add the current form values as URL parameters to the link instead.

This partially reverts commit 8d5f428c3e, but most of the changes it
introduced are left intact: either we’ll need them later (specifically
the overwrite flag of add_form_data_to_template()), or they don’t matter
anymore and might as well be left alone (such as the move of the
duplicates warning inside the form).

Part of #11.