Preserve form inputs in duplicate edit link (JS)
To complement the previous commit, which adds ?form_representation= to
the edit links for potential duplicates in no-JS mode, do the same in
JavaScript as soon as the link is clicked (or middle-clicked: “auxclick”
event). We clear any pre-existing form_representation parameters (mainly
in case the handler runs multiple times, in case of am middle-click),
but preserve others – currently we don’t expect any parameters to exist,
but maybe the tool will add some in the future, who knows.

Part of #11.