Fix current_url() for URLs with space
In the referrer, spaces are percent-encoded, but url_for() uses ‘+’
characters instead. There doesn’t seem to be a way to influence this,
but a replace() at the end seems to do the trick.

This becomes especially important when the current URL and referrer may
contain form_representation parameters, now that we no longer stash them
away (see previous commit).