Pop OAuth redirect target after redirect
When we send the user to log in via OAuth, we put the current location
into the session so that we can redirect them there once they return.
However, we never removed that URL from the session again, so it would
just be sent back and forth in the session cookie all the time. That’s
wasteful and completely unnecessary – just remove it when we don’t need
it anymore.