Remove code to fix session token
When this code was originally added, in commit e5134ed087 of December
2018, I guessed that it could be removed “after, I don’t know, half a
year?” Needless to say, that didn’t happen. Even the move to
toolforge.org (commit 0dadbb4d4e of April 2020), which *definitely* made
this code obsolete, was more than half a year ago. In other words, it’s
high time for this code to be removed.