Replace current_url() with flask.request.url
The old implementation doesn’t work properly if the request’s URL
parameters were encoded in an unusual way, e.g. with a trailing &.
Moreover, as far as I can tell it’s *completely unnecessary*:
flask.request.url does the job just as well. I can only assume that when
I wrote (and rewrote, and tweaked) this function, I was not aware of the
url member. (The request object is a bit hard to explore because in the
in-browser debugger it’s no longer available; to avoid this, stash
flask.request._get_current_object() somewhere to resolve the proxy while
the request context is alive.)

We keep the function around for now because some of the tests rely on
monkeypatching it, but otherwise we might just as well simply inline it.

The include_args argument had already been unused since the “stash form
data” mechanism was removed in commit b5fb43bb14.