Export other messages to MediaWiki format
Remove the code that tracks and updates the variables and lists, and
instead hard-code the values that the previous version printed. (One
more variable name had to be added by hand: the bulk_not_allowed message
only uses the 'user' variable in the Italian translation.) Then, write
out the MediaWiki versions in all languages. (Including English, but
that file doesn’t change as a result.)