Don’t write duplicate messages to i18n/
Internally, we have several copies of a few messages that are used in
different places, sometimes in different capitalization. I never put
these extra messages on the wiki pages where translation was done so far
– I’d just lower- or uppercase the initial word as necessary (using
Emacs to handle Unicode capitalization) when transcribing the
translations. Do something similar here: have a dict of messages that
can be reconstructed from other messages using a transformation
function, and when encountering those messages, assert that they match
the transformed source message instead of writing them out to i18n/.
This turns out to highlight two incorrectly capitalized messages, too,
so those are fixed in this commit.