Add function to load translations from i18n/
And a test to check that the loaded translations are identical to the
originals, i.e. that converting from Python syntax to MediaWiki syntax
and back is a lossless roundtrip. The test is a pytest, rather than a
plain assert in translations.py, mainly so that the assertion failure
prints a nice error message if the assertion fails – this was useful to
figure out some issues (see also the previous two commits). Now that the
assertion passes, it’s not so important anymore.