Load translations from i18n/
This is probably the last step in the translatewiki.net migration as far
as this tool’s source code is concerned (the rest of the migration will
presumably be to make translatewiki.net import the translations and then
periodically reexport them). The Python versions of the translations are
removed, now that we’re satisfied they can be completely reconstructed
from the MediaWiki versions. The translations are eagerly loaded as
soon as the module is imported – I don’t see any point in delaying it,
since we’ll probably need all the translations anyways (this isn’t PHP:
one Python process serves many requests, for different languages).

The py2mw function is kept for now, though it’s currently unused. I’ll
probably discard it at some point.